+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 
 # 2.2.2
 * bugfix KeywordExtractor: Also put Character Class M,Cs,Co to L
+* bugfix KeywordExtractor: works with upper, part and alpha
 
 ## 2.2.1
 * bugfix Keyword extractor